'Final' tuning of the bot drivetrain after lock-tite and grinding off sharp gears.
Added smarter vision code to help identify the box of interest on the frame for our cludgy color-based vision.
This uses tensorFlow to get a rough box of the ring 4-stack (this works pretty reliably.)
Then the gamepad can be used to fine-tune the starting box from tensor flow to be more exact.