Delete unused MiniPID file -- we're using the REV PID in the REV Hub.
Update the autonomous into 3 different sections (one per ring count.)
Inside those functions, after every significant operation, check if the opMode is still active.  If not active, return.  This is to handle the fact that the OpMode has about 250ms to return after stop is pressed or it throws an error.