Switch the shooter over from the REV Hub PID to a SW PID. We're trying to understand wny the ring shooting seems a little wild. When gathering and visualizing the encoder counts, it seems like there are delays that cause some readings to have higher or lower counts.  We can hear the REV hub changing RPM slightly in response (we think.)
The switch to a SW PID was to provide some filtering of the input data.
It's a time-weighted PID targetting a 50ms sample rate.  The samples are collected into a 10-entry circular buffer so in 'steady-state' the sample window is 500ms.  When the instantaneous RPM drops below 90% of target, the PID switches to a single-sample mode until the target is reached (and then another 500ms).  This reduces the noise in steady-state we were seeing also eliminates the lag problem for excursions like startup and shooting.

All the Autos were retuned for an 'RPM' value instead of REV HUB power value.
'F' for the SW PID is battery voltage compensated.

To support SW PID, all the drive commands were converted to 'Async' and a routine was written to monitor PID, Drive updates, and sleep requests.  (CheckWait)

Brought back the 'MiniPID' module and added support for time-scaling of 'I' and 'D' factors.

In theory, setting 'SWPID' to false should switch everything back to the REV Hub PID and power settings.  This path is completely untested.